pyproject.toml
[build-system]
requires = [ "setuptools>=61.0" ]
build-backend = "setuptools.build_meta"

[project]
name = "hatch-xclam"
version = "0.8.0-dev.2"
description = "Package manager for the Cracking Shells ecosystem"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
  "Programming Language :: Python :: 3.12",
  "License :: OSI Approved :: GNU Affero General Public License v3",
  "Operating System :: OS Independent"
]
dependencies = [
  "jsonschema>=4.0.0",
  "requests>=2.25.0",
  "packaging>=20.0",
  "docker>=7.1.0",
  "pydantic>=2.0.0",
  "hatch-validator>=0.8.0",
  "tomli-w>=1.0.0"
]

  [[project.authors]]
  name = "Cracking Shells Team"

  [project.optional-dependencies]
  docs = [ "mkdocs>=1.4.0", "mkdocstrings[python]>=0.20.0" ]
  dev = [
  "cs-wobble>=0.2.0",
  "pytest>=8.0.0",
  "black>=23.0.0",
  "ruff>=0.1.9",
  "pre-commit>=3.0.0"
]

  [project.scripts]
  hatch = "hatch.cli:main"

  [project.urls]
  Homepage = "https://github.com/CrackingShells/Hatch"
  "Bug Tracker" = "https://github.com/CrackingShells/Hatch/issues"

[tool.setuptools.package-dir]
"" = "."

[tool.setuptools.packages.find]
where = [ "." ]

[tool.pytest.ini_options]
pythonpath = [ "tests" ]

[tool.ruff]
exclude = [
  ".git",
  ".ruff_cache",
  "__pycache__",
  "node_modules",
  "site",
  "py-repo-template",
  "cracking-shells-playbook",
  "__reports__",
  "__design__",
  "tests/test_data"
]

  [tool.ruff.lint]
  select = [ "E", "F" ]
  ignore = [ "E501" ]
